requirements.txt
python-dotenv
pyrogram>=2.0
tgcrypto
motor
dnspython